+++ NEWS.md
@@ -1,5 +1,6 @@
 # ppseq (development version)
 
+* added `calc_next()` function to calculate probability of response in next patient (Closed Issue #2)
 * fixed a bug in the default arguments to `optimize_design()` (Closed Issue #12)
 * added efficacy monitoring option, default is still futility monitoring (Closed Issue #3)